Added strict filter for urls from render context menu
