+++ android/android_browser_tests.gni
@@ -48,12 +48,6 @@ android_test_exception_sources = [
   "//brave/browser/extensions/brave_theme_event_router_browsertest.cc",
   "//brave/browser/extensions/extension_urls_browsertest.cc",
   "//brave/browser/favicon/favicon_database_browsertest.cc",
-  "//brave/browser/net/brave_accept_header_browsertest.cc",
-  "//brave/browser/net/brave_network_delegate_browsertest.cc",
-  "//brave/browser/net/brave_network_delegate_hsts_fingerprinting_browsertest.cc",
-  "//brave/browser/net/brave_service_key_network_delegate_helper_browsertest.cc",
-  "//brave/browser/net/brave_site_hacks_network_delegate_helper_browsertest.cc",
-  "//brave/browser/net/global_privacy_control_network_delegate_helper_browsertest.cc",
   "//brave/browser/policy/brave_policy_browsertest.cc",
   "//brave/browser/profiles/brave_bookmark_model_loaded_observer_browsertest.cc",
   "//brave/browser/renderer_context_menu/brave_mock_render_view_context_menu.cc",
@@ -126,6 +120,7 @@ android_test_exception_deps = [
   "//brave/browser/decentralized_dns/test:browser_tests",
   "//brave/browser/extensions",
   "//brave/browser/farbling:browser_tests",
+  "//brave/browser/net:browser_tests",
   "//brave/browser/permissions:browser_tests",
   "//brave/browser/test:browser_tests",
   "//brave/browser/tor:browser_tests",
