+++ browser/net/BUILD.gn
@@ -35,6 +35,8 @@ source_set("net") {
     "brave_proxying_url_loader_factory.h",
     "brave_proxying_web_socket.cc",
     "brave_proxying_web_socket.h",
+    "brave_query_filter.cc",
+    "brave_query_filter.h",
     "brave_reduce_language_network_delegate_helper.cc",
     "brave_reduce_language_network_delegate_helper.h",
     "brave_request_handler.cc",
@@ -128,5 +130,93 @@ source_set("net") {
       "//brave/components/brave_webtorrent/browser/",
       "//extensions/browser",
     ]
+  }
+}
+
+source_set("browser_tests") {
+  testonly = true
+
+  sources = [
+    "brave_accept_header_browsertest.cc",
+    "brave_network_delegate_browsertest.cc",
+    "brave_network_delegate_hsts_fingerprinting_browsertest.cc",
+    "brave_service_key_network_delegate_helper_browsertest.cc",
+    "brave_site_hacks_network_delegate_helper_browsertest.cc",
+    "brave_system_request_handler_browsertest.cc",
+    "global_privacy_control_network_delegate_helper_browsertest.cc",
+  ]
+  deps = [
+    "//brave/browser/net",
+    "//brave/components/brave_shields/browser",
+    "//brave/components/brave_shields/common",
+    "//brave/components/tor",
+    "//chrome/browser",
+    "//chrome/browser:browser_process",
+    "//chrome/browser/profiles:profile",
+    "//chrome/browser/ui",
+    "//components/content_settings/core/browser",
+    "//components/prefs",
+    "//services/network/public/cpp",
+  ]
+  if (!is_android) {
+    deps += [ "//chrome/test:test_support_ui" ]
+  }
+  if (is_android) {
+    deps += [
+      "//chrome/test:test_support_ui_android",
+      "//net:test_support",
+      "//net/traffic_annotation:test_support",
+    ]
+  }
+  defines = [ "HAS_OUT_OF_PROC_TEST_RUNNER" ]
+}
+
+source_set("unit_tests") {
+  testonly = true
+
+  sources = [
+    "brave_ad_block_tp_network_delegate_helper_unittest.cc",
+    "brave_block_safebrowsing_urls_unittest.cc",
+    "brave_common_static_redirect_network_delegate_helper_unittest.cc",
+    "brave_httpse_network_delegate_helper_unittest.cc",
+    "brave_network_delegate_base_unittest.cc",
+    "brave_query_filter_unittest.cc",
+    "brave_site_hacks_network_delegate_helper_unittest.cc",
+    "brave_static_redirect_network_delegate_helper_unittest.cc",
+    "brave_system_request_handler_unittest.cc",
+  ]
+
+  deps = [
+    "//brave/browser/net",
+    "//brave/browser/net:geolocation",
+    "//brave/components/brave_shields/browser",
+    "//chrome/test:test_support",
+    "//content/test:test_support",
+    "//net",
+    "//net:test_support",
+    "//services/network:network_service",
+    "//testing/gmock",
+    "//testing/gtest",
+    "//third_party/abseil-cpp:absl",
+  ]
+
+  if (enable_brave_webtorrent) {
+    sources += [
+      "//brave/browser/extensions/brave_webtorrent_navigation_throttle_unittest.cc",
+      "//brave/browser/net/brave_torrent_redirect_network_delegate_helper_unittest.cc",
+    ]
+
+    deps += [
+      "//brave/browser/extensions",
+      "//brave/components/brave_referrals/browser",
+    ]
+  }
+  if (enable_brave_referrals) {
+    if (!is_android) {
+      sources += [
+        # TODO(samartnik): this should work on Android, we will review it once unit tests are set up on CI
+        "brave_referrals_network_delegate_helper_unittest.cc",
+      ]
+    }
   }
 }
